requirements.txt
accelerate==0.33.0
datasets==2.19.2
fire==0.6.0
h5py==3.12.1
hydra-core==1.3.2
lightning==2.2.5
matplotlib==3.9.2
multipledispatch==1.0.0
omegaconf==2.3.0
pandas==2.2.3
peft==0.14.0
Pillow==10.4.0
scikit-learn
scipy
setuptools==69.5.1
torch==2.4.0
torchmetrics==1.4.0.post0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.44.0
wandb==0.18.3
rich
sentencepiece
deepspeed==0.15.4 